+++ cfg/hgcIdTuples.yaml
@@ -1,55 +1,24 @@
 
 common:
   output_dir:
+    # dictionary machine name -> directory to support different paths
     default: /eos/user/c/cerminar/hgcal/CMSSW1015/plots/
-    matterhorn: /Users/cerminar/cernbox/hgcal/CMSSW1015/plots/
-    Matterhorn: /Users/cerminar/cernbox/hgcal/CMSSW1015/plots/
     triolet: /Users/cerminar/cernbox/hgcal/CMSSW1015/plots/
-  output_dir_local: /Users/cerminar/cernbox/hgcal/CMSSW1015/plots/
-  output_dir_lx: /eos/user/c/cerminar/hgcal/CMSSW1015/plots/
   plot_version: v160A
   run_clustering: False
   run_density_computation: False
 # +AccountingGroup = "group_u_CMS.u_zh.users"
 # +AccountingGroup = "group_u_CMST3.all"
+  file_label: compid
 
-collections:
-      
-  egid:
-    file_label:
-      compid
-    samples:
-      # - ele_flat2to100_PU0
-      # - ele_flat2to100_PU200
-      # - doubleele_flat1to100_PU0
-      - doublephoton_flat1to100_PU200
-      - doubleele_flat1to100_PU200
-      # - nugun_alleta_pu200
-      # - photon_flat8to150_PU0
-      # - photon_flat8to150_PU200
-      # - dyll_PU200
-    plotters:
-      - !!python/name:cfg.hgcIdTuples.egid_plotters
-    htc_jobflavor:
-      microcentury
+egid:
+  plotters:
+    - !!python/name:cfg.hgcIdTuples.egid_plotters
 
-  puid:
-    file_label:
-      compid
-    samples:
-      - nugun_alleta_pu200
-    plotters:
-      - !!python/name:cfg.hgcIdTuples.pu_plotters
-    htc_jobflavor:
-      microcentury
+puid:
+  plotters:
+    - !!python/name:cfg.hgcIdTuples.pu_plotters
 
-  piid:
-    file_label:
-      compid
-    samples:
-      - ttbar_PU200
-    plotters:
-      - !!python/name:cfg.hgcIdTuples.piid_plotters
-    htc_jobflavor:
-      microcentury
-      
+piid:
+  plotters:
+    - !!python/name:cfg.hgcIdTuples.piid_plotters
